requirements.txt
fastapi
python-dotenv
google-generativeai
pydantic
